% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif_acc.R
\name{acc}
\alias{acc}
\title{Classification Accuracy}
\usage{
acc(truth, response, ...)
}
\arguments{
\item{truth}{:: \code{factor()}\cr
True (observed) labels.
Must have the same levels and length as \code{response}.}

\item{response}{:: \code{factor()}\cr
Predicted response labels.
Must have the same levels and length as \code{truth}.}

\item{...}{:: any\cr
Additional arguments. Currently ignored.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Classification measure defined as \deqn{
  \frac{1}{n} \sum_{i=1}^n \left( t_i = r_i \right).
}{
  mean(t == r).
}
}
\section{Meta Information}{

\itemize{
\item Minimum value: 0
\item Maximum value: 1
\item Minimize: FALSE
}
}

\seealso{
Other Classification Measures: \code{\link{ce}}
}
\concept{Classification Measures}
