% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif_ber.R
\name{ber}
\alias{ber}
\title{Balanced Error Rate}
\usage{
ber(truth, response, ...)
}
\arguments{
\item{truth}{:: \code{factor()}\cr
True (observed) labels.
Must have the same levels and length as \code{response}.}

\item{response}{:: \code{factor()}\cr
Predicted response labels.
Must have the same levels and length as \code{truth}.}

\item{...}{:: any\cr
Additional arguments. Currently ignored.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Mean of misclassification error rates on all individual classes.

Classification measure defined as \deqn{
  \frac{1}{n} \sum_{i=1}^n \left( t_i \neq r_i \right).
}{
  mean(t != r).
}
}
\section{Meta Information}{

\itemize{
\item Minimum value: 0
\item Maximum value: 1
\item Minimize: TRUE
}
}

\seealso{
Other Classification Measures: \code{\link{acc}},
  \code{\link{ce}}, \code{\link{logloss}}
}
\concept{Classification Measures}
\concept{classification_measure}
