% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif_ce.R
\name{ce}
\alias{ce}
\title{Classification Error}
\usage{
ce(truth, response)
}
\arguments{
\item{truth}{:: \code{factor()}\cr
True (observed) labels.
Must have the same length as \code{response}.}

\item{response}{:: \code{factor()}\cr
Predicted response labels.
Must have the same length as \code{truth}.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Definition:
\deqn{
  \frac{1}{n} \sum_{i=1}^n \left( t_i \neq r_i \right).
}{
  mean(t != r)
}
}
\section{Meta Information}{

\itemize{
\item Minimum value: 0
\item Maximum value: 1
\item Minimize: TRUE
}
}

