% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_fbeta_score.R
\name{fbeta_score}
\alias{fbeta_score}
\title{F-beta Score}
\usage{
fbeta_score(truth, response, positive, beta = 1, na_value = NaN, ...)
}
\arguments{
\item{truth}{:: \code{factor()}\cr
True (observed) labels.
Must have exactly same two levels and the same length as \code{response}.}

\item{response}{:: \code{factor()}\cr
Predicted response labels.
Must have exactly same two levels and the same length as \code{truth}.}

\item{positive}{:: \code{character(1)}\cr
Name of the positive class.}

\item{beta}{:: \code{numeric(1)}\cr
Parameter to give either precision or recall more weight.
Default is 1, resulting in balanced weights.}

\item{na_value}{:: \code{numeric(1)}\cr
Value that should be returned if the measure is not defined for the input
(as described in the note). Default is \code{NaN}.}

\item{...}{:: any\cr
Additional arguments. Currently ignored.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Binary classification measure defined with \eqn{P} as \link{precision} and \eqn{R} as
\link{recall} as \deqn{
   (1 + \beta^2) \frac{PR}{(\beta^2 P) + R}.
}{
   (1 + beta^2) * (P*R) / ((beta^2 * P) + R).
}
For \eqn{\beta = 1}{beta = 1}, this is called "F1" score.
}
\note{
This measure is undefined if either \link{precision} or \link{recall} is undefined, i.e.
TP + FP = 0 or TP + FN = 0.
}
\section{Meta Information}{

\itemize{
\item Minimum value: 0
\item Maximum value: 1
\item Minimize: FALSE
}
}

\references{
\url{https://en.wikipedia.org/wiki/Template:DiagnosticTesting_Diagram}
}
\seealso{
Other Binary Classification Measures: \code{\link{auc}},
  \code{\link{dor}}, \code{\link{fdr}}, \code{\link{fnr}},
  \code{\link{fn}}, \code{\link{fomr}}, \code{\link{fpr}},
  \code{\link{fp}}, \code{\link{mcc}}, \code{\link{npv}},
  \code{\link{ppv}}, \code{\link{tnr}}, \code{\link{tn}},
  \code{\link{tpr}}, \code{\link{tp}}
}
\concept{Binary Classification Measures}
