% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_fomr.R
\name{fomr}
\alias{fomr}
\title{False Omission Rate}
\usage{
fomr(truth, response, positive, sample_weights = NULL, na_value = NaN, ...)
}
\arguments{
\item{truth}{(\code{factor()})\cr
True (observed) labels.
Must have the exactly same two levels and the same length as \code{response}.}

\item{response}{(\code{factor()})\cr
Predicted response labels.
Must have the exactly same two levels and the same length as \code{truth}.}

\item{positive}{(\verb{character(1))}\cr
Name of the positive class.}

\item{sample_weights}{(\code{numeric()})\cr
Vector of non-negative and finite sample weights.
Must have the same length as \code{truth}.
The vector gets automatically normalized to sum to one.
Defaults to equal sample weights.}

\item{na_value}{(\code{numeric(1)})\cr
Value that should be returned if the measure is not defined for the input
(as described in the note). Default is \code{NaN}.}

\item{...}{(\code{any})\cr
Additional arguments. Currently ignored.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Measure to compare true observed labels with predicted
labels
in binary classification tasks.
}
\details{
The False Omission Rate is defined as \deqn{
   \frac{\mathrm{FN}}{\mathrm{FN} + \mathrm{TN}}.
}{
   FN / (FN + TN).
}

This measure is undefined if FN + TN = 0.
}
\section{Meta Information}{

\itemize{
\item Type: \code{"binary"}
\item Range: \eqn{[0, 1]}{[0, 1]}
\item Minimize: \code{TRUE}
\item Required prediction: \code{response}
}
}

\examples{
set.seed(1)
lvls = c("a", "b")
truth = factor(sample(lvls, 10, replace = TRUE), levels = lvls)
response = factor(sample(lvls, 10, replace = TRUE), levels = lvls)
fomr(truth, response, positive = "a")
}
\references{
\url{https://en.wikipedia.org/wiki/Template:DiagnosticTesting_Diagram}
}
\seealso{
Other Binary Classification Measures:
\code{\link[=auc]{auc()}},
\code{\link[=bbrier]{bbrier()}},
\code{\link[=dor]{dor()}},
\code{\link[=fbeta]{fbeta()}},
\code{\link[=fdr]{fdr()}},
\code{\link[=fn]{fn()}},
\code{\link[=fnr]{fnr()}},
\code{\link[=fp]{fp()}},
\code{\link[=fpr]{fpr()}},
\code{\link[=gmean]{gmean()}},
\code{\link[=gpr]{gpr()}},
\code{\link[=npv]{npv()}},
\code{\link[=ppv]{ppv()}},
\code{\link[=prauc]{prauc()}},
\code{\link[=tn]{tn()}},
\code{\link[=tnr]{tnr()}},
\code{\link[=tp]{tp()}},
\code{\link[=tpr]{tpr()}}
}
\concept{Binary Classification Measures}
\concept{binary_classification_measure}
