% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif_fp.R
\name{fp}
\alias{fp}
\title{False Positives}
\usage{
fp(truth, response, positive)
}
\arguments{
\item{truth}{:: \code{factor()}\cr
True (observed) labels.
Must have exactly same two levels and the same length as \code{response}.}

\item{response}{:: \code{factor()}\cr
Predicted response labels.
Must have exactly same two levels and the same length as \code{response}.}

\item{positive}{:: \code{character(1)}\cr
Name of the positive class.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Classification measure counting the false positives.
}
\section{Meta Information}{

\itemize{
\item Minimum value: 0
\item Maximum value: Inf
\item Minimize: FALSE
}
}

\references{
\url{https://en.wikipedia.org/wiki/Template:DiagnosticTesting_Diagram}
}
