% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr_kendalls_tau.R
\name{kendalls_tau}
\alias{kendalls_tau}
\title{Kendall's tau}
\usage{
kendalls_tau(truth, response)
}
\arguments{
\item{truth}{:: \code{numeric()}\cr
True (observed) values.
Must have the same length as \code{response}.}

\item{response}{:: \code{numeric()}\cr
Predicted response values.
Must have the same length as \code{truth}.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Defined as Kendall rank correlation coefficent between truth and response.
Calls \code{\link[stats:cor]{stats::cor()}} with \code{method} set to \code{"kendall"}.
}
\section{Meta Information}{

\itemize{
\item Minimum value: -Inf
\item Maximum value: 1
\item Minimize: FALSE
}
}

\references{
\cite{rosset_2006}
}
