% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr_ktau.R
\name{ktau}
\alias{ktau}
\title{Kendall's tau}
\usage{
ktau(truth, response, ...)
}
\arguments{
\item{truth}{(\code{numeric()})\cr
True (observed) values.
Must have the same length as \code{response}.}

\item{response}{(\code{numeric()})\cr
Predicted response values.
Must have the same length as \code{truth}.}

\item{...}{(\code{any})\cr
Additional arguments. Currently ignored.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Measure to compare true observed response with predicted response in regression tasks.
}
\details{
Kendall's tau is defined as Kendall's rank correlation coefficient between truth and response.
It is defined as \deqn{
  \tau = \frac{(\mathrm{number of concordant pairs)} -
  (\mathrm{number of discordant pairs)}}{\mathrm{(number of pairs)}}
  }{
    t = (number of concordant pairs) - (number of discordant pairs) / (number of pairs)}
Calls \code{\link[stats:cor]{stats::cor()}} with \code{method} set to \code{"kendall"}.
}
\section{Meta Information}{

\itemize{
\item Type: \code{"regr"}
\item Range: \eqn{[-1, 1]}{[-1, 1]}
\item Minimize: \code{FALSE}
\item Required prediction: \code{response}
}
}

\examples{
set.seed(1)
truth = 1:10
response = truth + rnorm(10)
ktau(truth, response)
}
\references{
Rosset S, Perlich C, Zadrozny B (2006).
\dQuote{Ranking-based evaluation of regression models.}
\emph{Knowledge and Information Systems}, \bold{12}(3), 331--353.
\doi{10.1007/s10115-006-0037-3}.
}
\seealso{
Other Regression Measures:
\code{\link[=ae]{ae()}},
\code{\link[=ape]{ape()}},
\code{\link[=bias]{bias()}},
\code{\link[=linex]{linex()}},
\code{\link[=mae]{mae()}},
\code{\link[=mape]{mape()}},
\code{\link[=maxae]{maxae()}},
\code{\link[=maxse]{maxse()}},
\code{\link[=medae]{medae()}},
\code{\link[=medse]{medse()}},
\code{\link[=mse]{mse()}},
\code{\link[=msle]{msle()}},
\code{\link[=pbias]{pbias()}},
\code{\link[=pinball]{pinball()}},
\code{\link[=rae]{rae()}},
\code{\link[=rmse]{rmse()}},
\code{\link[=rmsle]{rmsle()}},
\code{\link[=rrse]{rrse()}},
\code{\link[=rse]{rse()}},
\code{\link[=rsq]{rsq()}},
\code{\link[=sae]{sae()}},
\code{\link[=se]{se()}},
\code{\link[=sle]{sle()}},
\code{\link[=smape]{smape()}},
\code{\link[=srho]{srho()}},
\code{\link[=sse]{sse()}}
}
\concept{Regression Measures}
\concept{regression_measure}
