% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr_linex.R
\name{linex}
\alias{linex}
\title{Linear-Exponential Loss (per observation)}
\usage{
linex(truth, response, a = -1, b = 1, ...)
}
\arguments{
\item{truth}{(\code{numeric()})\cr
True (observed) values.
Must have the same length as \code{response}.}

\item{response}{(\code{numeric()})\cr
Predicted response values.
Must have the same length as \code{truth}.}

\item{a}{(\code{numeric(1)})\cr
Shape parameter controlling asymmetry.
Negative values penalize overestimation more, positive values penalize underestimation more.
As \code{a} approaches 0, the loss resembles squared error loss. Default is \code{-1}.}

\item{b}{(\code{numeric(1)})\cr
Positive scaling factor for the loss. Larger values increase the loss magnitude.
Default is \code{1}.}

\item{...}{(\code{any})\cr
Additional arguments. Currently ignored.}
}
\value{
Performance value as \code{numeric(length(truth))}.
}
\description{
Measure to compare true observed response with predicted response in regression tasks.

Note that this is an unaggregated measure, returning the losses per observation.
}
\details{
The Linear-Exponential Loss is defined as \deqn{
  b (\exp (t_i - r_i) - a (t_i - r_i) - 1),
}{
}
where \eqn{a \neq 0, b > 0}.
}
\section{Meta Information}{

\itemize{
\item Type: \code{"regr"}
\item Range (per observation): \eqn{[0, \infty)}{[0, Inf)}
\item Minimize (per observation): \code{TRUE}
\item Required prediction: \code{response}
}
}

\examples{
set.seed(1)
truth = 1:10
response = truth + rnorm(10)
linex(truth, response)
}
\references{
Varian, R. H (1975).
\dQuote{A Bayesian Approach to Real Estate Assessment.}
In Fienberg SE, Zellner A (eds.), \emph{Studies in Bayesian Econometrics and Statistics: In Honor of Leonard J. Savage}, 195--208.
North-Holland, Amsterdam.
}
\seealso{
Other Regression Measures:
\code{\link[=ae]{ae()}},
\code{\link[=ape]{ape()}},
\code{\link[=bias]{bias()}},
\code{\link[=ktau]{ktau()}},
\code{\link[=mae]{mae()}},
\code{\link[=mape]{mape()}},
\code{\link[=maxae]{maxae()}},
\code{\link[=maxse]{maxse()}},
\code{\link[=medae]{medae()}},
\code{\link[=medse]{medse()}},
\code{\link[=mse]{mse()}},
\code{\link[=msle]{msle()}},
\code{\link[=pbias]{pbias()}},
\code{\link[=pinball]{pinball()}},
\code{\link[=rae]{rae()}},
\code{\link[=rmse]{rmse()}},
\code{\link[=rmsle]{rmsle()}},
\code{\link[=rrse]{rrse()}},
\code{\link[=rse]{rse()}},
\code{\link[=rsq]{rsq()}},
\code{\link[=sae]{sae()}},
\code{\link[=se]{se()}},
\code{\link[=sle]{sle()}},
\code{\link[=smape]{smape()}},
\code{\link[=srho]{srho()}},
\code{\link[=sse]{sse()}}
}
\concept{Regression Measures}
\concept{regression_measure}
