% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr_mae.R
\name{mae}
\alias{mae}
\title{Mean Absolute Errors}
\usage{
mae(truth, response)
}
\arguments{
\item{truth}{:: \code{numeric()}\cr
True (observed) values.
Must have the same length as \code{response}.}

\item{response}{:: \code{numeric()}\cr
Predicted response values.
Must have the same length as \code{truth}.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Definition:
\deqn{
  \frac{1}{n} \sum_{i=1}^n \left| t_i - r_i \right|.
}{
  mean(abs(t - r))
}
}
\section{Meta Information}{

\itemize{
\item Minimum value: 0
\item Maximum value: Inf
\item Minimize: TRUE
}
}

