% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr_mape.R
\name{mape}
\alias{mape}
\title{Mean Absolute Percent Error}
\usage{
mape(truth, response, na_value = NaN)
}
\arguments{
\item{truth}{:: \code{numeric()}\cr
True (observed) values.
Must have the same length as \code{response}.}

\item{response}{:: \code{numeric()}\cr
Predicted response values.
Must have the same length as \code{truth}.}

\item{na_value}{:: \code{numeric(1)}\cr
Value that should be returned if the measure is not defined for the input
(as described in the note). Default is \code{NaN}.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Definition:
\deqn{
  \frac{1}{n} \sum_{i=1}^n \frac{\left| t_i - r_i \right|}{t_i}.
 }{
   mean(abs(t - r) / t)
 }
}
\note{
This measure is undefined if any element of \eqn{t} is \eqn{0}.
}
\section{Meta Information}{

\itemize{
\item Minimum value: 0
\item Maximum value: Inf
\item Minimize: TRUE
}
}

\references{
\cite{de_myttenaere_2016}
}
