% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr_maxse.R
\name{maxse}
\alias{maxse}
\title{Max Squared Error}
\usage{
maxse(truth, response)
}
\arguments{
\item{truth}{:: \code{numeric()}\cr
True (observed) values.
Must have the same length as \code{response}.}

\item{response}{:: \code{numeric()}\cr
Predicted response values.
Must have the same length as \code{truth}.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Regression measure defined as \deqn{
  \max \left( t_i - r_i \right)^2.
 }{
   max((t - r)^2).
 }
}
\section{Meta Information}{

\itemize{
\item Minimum value: 0
\item Maximum value: Inf
\item Minimize: TRUE
}
}

