% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{measures}
\alias{measures}
\alias{list_measures}
\title{Measure Registry}
\format{An object of class \code{environment} of length 39.}
\usage{
measures

list_measures()
}
\description{
The \code{\link[=environment]{environment()}} \code{measures} keeps track of all measures in this package.
Stores meta information about measures, such as minimum, maximum of if the
measure must be minimized or maximized.

\code{list_measures()} returns the information in as a \code{\link[=data.frame]{data.frame()}}.
}
\keyword{datasets}
