% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr_msle.R
\name{msle}
\alias{msle}
\title{Mean Squared Log Error}
\usage{
msle(truth, response, na_value = NaN, ...)
}
\arguments{
\item{truth}{:: \code{numeric()}\cr
True (observed) values.
Must have the same length as \code{response}.}

\item{response}{:: \code{numeric()}\cr
Predicted response values.
Must have the same length as \code{truth}.}

\item{na_value}{:: \code{numeric(1)}\cr
Value that should be returned if the measure is not defined for the input
(as described in the note). Default is \code{NaN}.}

\item{...}{:: \code{any}\cr
Additional arguments. Currently ignored.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Regression measure defined as \deqn{
  \frac{1}{n} \sum_{i=1}^n \left( \ln (1 + t_i) - \ln (1 + r_i) \right)^2.
}{
  mean(log(1 + t) - log(1 + r))^2.
}
}
\note{
This measure is undefined if any element of \eqn{t} or \eqn{r} is less than or equal to \eqn{-1}.
}
\section{Meta Information}{

\itemize{
\item Type: \code{"regr"}
\item Range: \eqn{[0, \infty)}{[0, Inf)}
\item Minimize: \code{TRUE}
\item Required prediction: \code{response}
}
}

\examples{
set.seed(1)
truth = 1:10
response = truth + rnorm(10)
msle(truth, response)
}
\seealso{
Other Regression Measures: \code{\link{bias}},
  \code{\link{ktau}}, \code{\link{mae}},
  \code{\link{mape}}, \code{\link{maxae}},
  \code{\link{maxse}}, \code{\link{medae}},
  \code{\link{medse}}, \code{\link{mse}},
  \code{\link{pbias}}, \code{\link{rae}},
  \code{\link{rmse}}, \code{\link{rmsle}},
  \code{\link{rrse}}, \code{\link{rse}}, \code{\link{rsq}},
  \code{\link{sae}}, \code{\link{smape}},
  \code{\link{srho}}, \code{\link{sse}}
}
\concept{Regression Measures}
\concept{regression_measure}
