% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif_logloss.R
\name{obs_logloss}
\alias{obs_logloss}
\title{Observation-wise Log Loss}
\usage{
obs_logloss(truth, prob)
}
\description{
Observation-wise loss function for \code{\link[=logloss]{logloss()}}.
}
\keyword{internal}
