% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_prauc.R
\name{prauc}
\alias{prauc}
\title{Area Under the Precision-Recall Curve}
\usage{
prauc(truth, prob, positive, na_value = NaN, ...)
}
\arguments{
\item{truth}{(\code{factor()})\cr
True (observed) labels.
Must have the exactly same two levels and the same length as \code{response}.}

\item{prob}{(\code{numeric()})\cr
Predicted probability for positive class.
Must have exactly same length as \code{truth}.}

\item{positive}{(\verb{character(1))}\cr
Name of the positive class.}

\item{na_value}{(\code{numeric(1)})\cr
Value that should be returned if the measure is not defined for the input
(as described in the note). Default is \code{NaN}.}

\item{...}{(\code{any})\cr
Additional arguments. Currently ignored.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Measure to compare true observed labels with predicted
probabilities
in binary classification tasks.
}
\details{
Computes the area under the Precision-Recall curve (PRC).
The PRC can be interpreted as the relationship between precision and recall (sensitivity),
and is considered to be a more appropriate measure for unbalanced datasets than the ROC curve.
The AUC-PRC is computed by integration of the piecewise function.

This measure is undefined if the true values are either all positive or
all negative.
}
\section{Meta Information}{

\itemize{
\item Type: \code{"binary"}
\item Range: \eqn{[0, 1]}{[0, 1]}
\item Minimize: \code{FALSE}
\item Required prediction: \code{prob}
}
}

\examples{
truth = factor(c("a", "a", "a", "b"))
prob = c(.6, .7, .1, .4)
prauc(truth, prob, "a")
}
\references{
Davis J, Goadrich M (2006).
\dQuote{The relationship between precision-recall and ROC curves.}
In \emph{Proceedings of the 23rd International Conference on Machine Learning}.
ISBN 9781595933836.
}
\seealso{
Other Binary Classification Measures:
\code{\link[=auc]{auc()}},
\code{\link[=bbrier]{bbrier()}},
\code{\link[=dor]{dor()}},
\code{\link[=fbeta]{fbeta()}},
\code{\link[=fdr]{fdr()}},
\code{\link[=fn]{fn()}},
\code{\link[=fnr]{fnr()}},
\code{\link[=fomr]{fomr()}},
\code{\link[=fp]{fp()}},
\code{\link[=fpr]{fpr()}},
\code{\link[=gmean]{gmean()}},
\code{\link[=gpr]{gpr()}},
\code{\link[=npv]{npv()}},
\code{\link[=ppv]{ppv()}},
\code{\link[=tn]{tn()}},
\code{\link[=tnr]{tnr()}},
\code{\link[=tp]{tp()}},
\code{\link[=tpr]{tpr()}}
}
\concept{Binary Classification Measures}
\concept{binary_classification_measure}
