% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr_rae.R
\name{rae}
\alias{rae}
\title{Relative Absolute Error}
\usage{
rae(truth, response, na_value = NaN, ...)
}
\arguments{
\item{truth}{(\code{numeric()})\cr
True (observed) values.
Must have the same length as \code{response}.}

\item{response}{(\code{numeric()})\cr
Predicted response values.
Must have the same length as \code{truth}.}

\item{na_value}{(\code{numeric(1)})\cr
Value that should be returned if the measure is not defined for the input
(as described in the note). Default is \code{NaN}.}

\item{...}{(\code{any})\cr
Additional arguments. Currently ignored.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Measure to compare true observed response with predicted response in regression tasks.
}
\details{
The Relative Absolute Error is defined as \deqn{
  \frac{\sum_{i=1}^n \left| t_i - r_i \right|}{\sum_{i=1}^n \left| t_i - \bar{t} \right|},
}{
  sum((t - r)^2) / sum((t - mean(t))^2),
}
where \eqn{\bar{t} = \sum_{i=1}^n t_i}.
This measure is undefined for constant \eqn{t}.

Can be interpreted as absolute error of the predictions relative to a naive model predicting the mean.
}
\section{Meta Information}{

\itemize{
\item Type: \code{"regr"}
\item Range: \eqn{[0, \infty)}{[0, Inf)}
\item Minimize: \code{TRUE}
\item Required prediction: \code{response}
}
}

\examples{
set.seed(1)
truth = 1:10
response = truth + rnorm(10)
rae(truth, response)
}
\seealso{
Other Regression Measures:
\code{\link[=ae]{ae()}},
\code{\link[=ape]{ape()}},
\code{\link[=bias]{bias()}},
\code{\link[=ktau]{ktau()}},
\code{\link[=linex]{linex()}},
\code{\link[=mae]{mae()}},
\code{\link[=mape]{mape()}},
\code{\link[=maxae]{maxae()}},
\code{\link[=maxse]{maxse()}},
\code{\link[=medae]{medae()}},
\code{\link[=medse]{medse()}},
\code{\link[=mse]{mse()}},
\code{\link[=msle]{msle()}},
\code{\link[=pbias]{pbias()}},
\code{\link[=pinball]{pinball()}},
\code{\link[=rmse]{rmse()}},
\code{\link[=rmsle]{rmsle()}},
\code{\link[=rrse]{rrse()}},
\code{\link[=rse]{rse()}},
\code{\link[=rsq]{rsq()}},
\code{\link[=sae]{sae()}},
\code{\link[=se]{se()}},
\code{\link[=sle]{sle()}},
\code{\link[=smape]{smape()}},
\code{\link[=srho]{srho()}},
\code{\link[=sse]{sse()}}
}
\concept{Regression Measures}
\concept{regression_measure}
