% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{regr_params}
\alias{regr_params}
\title{Regression Parameters}
\arguments{
\item{truth}{:: \code{numeric()}\cr
True (observed) values.
Must have the same length as \code{response}.}

\item{response}{:: \code{numeric()}\cr
Predicted response values.
Must have the same length as \code{truth}.}

\item{...}{:: \code{any}\cr
Additional arguments. Currently ignored.}
}
\description{
Regression Parameters
}
\keyword{internal}
