% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr_rmse.R
\name{rmse}
\alias{rmse}
\title{Root Mean Squared Error}
\usage{
rmse(truth, response, sample_weights = NULL, ...)
}
\arguments{
\item{truth}{(\code{numeric()})\cr
True (observed) values.
Must have the same length as \code{response}.}

\item{response}{(\code{numeric()})\cr
Predicted response values.
Must have the same length as \code{truth}.}

\item{sample_weights}{(\code{numeric()})\cr
Vector of non-negative and finite sample weights.
Must have the same length as \code{truth}.
The vector gets automatically normalized to sum to one.
Defaults to equal sample weights.}

\item{...}{(\code{any})\cr
Additional arguments. Currently ignored.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Measure to compare true observed response with predicted response in regression tasks.
}
\details{
The Root Mean Squared Error is defined as \deqn{
  \sqrt{\frac{1}{n} \sum_{i=1}^n w_i \left( t_i - r_i \right)^2},
}{
  sqrt(weighted.mean((t - r)^2, w)),
}
where \eqn{w_i} are normalized sample weights.
}
\section{Meta Information}{

\itemize{
\item Type: \code{"regr"}
\item Range: \eqn{[0, \infty)}{[0, Inf)}
\item Minimize: \code{TRUE}
\item Required prediction: \code{response}
}
}

\examples{
set.seed(1)
truth = 1:10
response = truth + rnorm(10)
rmse(truth, response)
}
\seealso{
Other Regression Measures:
\code{\link[=ae]{ae()}},
\code{\link[=ape]{ape()}},
\code{\link[=bias]{bias()}},
\code{\link[=ktau]{ktau()}},
\code{\link[=linex]{linex()}},
\code{\link[=mae]{mae()}},
\code{\link[=mape]{mape()}},
\code{\link[=maxae]{maxae()}},
\code{\link[=maxse]{maxse()}},
\code{\link[=medae]{medae()}},
\code{\link[=medse]{medse()}},
\code{\link[=mse]{mse()}},
\code{\link[=msle]{msle()}},
\code{\link[=pbias]{pbias()}},
\code{\link[=pinball]{pinball()}},
\code{\link[=rae]{rae()}},
\code{\link[=rmsle]{rmsle()}},
\code{\link[=rrse]{rrse()}},
\code{\link[=rse]{rse()}},
\code{\link[=rsq]{rsq()}},
\code{\link[=sae]{sae()}},
\code{\link[=se]{se()}},
\code{\link[=sle]{sle()}},
\code{\link[=smape]{smape()}},
\code{\link[=srho]{srho()}},
\code{\link[=sse]{sse()}}
}
\concept{Regression Measures}
\concept{regression_measure}
