% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr_rsq.R
\name{rsq}
\alias{rsq}
\title{R Squared}
\usage{
rsq(truth, response, na_value = NaN, ...)
}
\arguments{
\item{truth}{(\code{numeric()})\cr
True (observed) values.
Must have the same length as \code{response}.}

\item{response}{(\code{numeric()})\cr
Predicted response values.
Must have the same length as \code{truth}.}

\item{na_value}{(\code{numeric(1)})\cr
Value that should be returned if the measure is not defined for the input
(as described in the note). Default is \code{NaN}.}

\item{...}{(\code{any})\cr
Additional arguments. Currently ignored.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Measure to compare true observed response with predicted response in regression tasks.
}
\details{
R Squared is defined as \deqn{
  1 - \frac{\sum_{i=1}^n \left( t_i - r_i \right)^2}{\sum_{i=1}^n \left( t_i - \bar{t} \right)^2},
}{
  1 - sum((t - r)^2) / sum((t - mean(t))^2),
}
where \eqn{\bar{t} = \sum_{i=1}^n t_i}.

Also known as coefficient of determination or explained variation.
Subtracts the \code{\link[=rse]{rse()}} from 1, hence it compares the squared error of
the predictions relative to a naive model predicting the mean.

This measure is undefined for constant \eqn{t}.
}
\section{Meta Information}{

\itemize{
\item Type: \code{"regr"}
\item Range: \eqn{(-\infty, 1]}{(-Inf, 1]}
\item Minimize: \code{FALSE}
\item Required prediction: \code{response}
}
}

\examples{
set.seed(1)
truth = 1:10
response = truth + rnorm(10)
rsq(truth, response)
}
\seealso{
Other Regression Measures:
\code{\link[=ae]{ae()}},
\code{\link[=ape]{ape()}},
\code{\link[=bias]{bias()}},
\code{\link[=ktau]{ktau()}},
\code{\link[=linex]{linex()}},
\code{\link[=mae]{mae()}},
\code{\link[=mape]{mape()}},
\code{\link[=maxae]{maxae()}},
\code{\link[=maxse]{maxse()}},
\code{\link[=medae]{medae()}},
\code{\link[=medse]{medse()}},
\code{\link[=mse]{mse()}},
\code{\link[=msle]{msle()}},
\code{\link[=pbias]{pbias()}},
\code{\link[=pinball]{pinball()}},
\code{\link[=rae]{rae()}},
\code{\link[=rmse]{rmse()}},
\code{\link[=rmsle]{rmsle()}},
\code{\link[=rrse]{rrse()}},
\code{\link[=rse]{rse()}},
\code{\link[=sae]{sae()}},
\code{\link[=se]{se()}},
\code{\link[=sle]{sle()}},
\code{\link[=smape]{smape()}},
\code{\link[=srho]{srho()}},
\code{\link[=sse]{sse()}}
}
\concept{Regression Measures}
\concept{regression_measure}
