% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_bbrier.R
\name{se_binary}
\alias{se_binary}
\title{Binary Squared Error}
\usage{
se_binary(truth, prob, positive, ...)
}
\description{
Observation-wise loss function for \code{\link[=bbrier]{bbrier()}}.
}
\keyword{internal}
