% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\name{similarity_params}
\alias{similarity_params}
\title{Similarity Parameters}
\arguments{
\item{sets}{(\code{list()})\cr
List of character or integer vectors.
\code{sets} must have at least 2 elements.}

\item{p}{(\code{integer(1)})\cr
Total number of possible elements.}

\item{na_value}{(\code{numeric(1)})\cr
Value that should be returned if the measure is not defined for the input
(as described in the note). Default is \code{NaN}.}

\item{...}{(\code{any})\cr
Additional arguments. Currently ignored.}
}
\description{
Similarity Parameters
}
\keyword{internal}
