% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{thresholds}
\alias{thresholds}
\title{Calculate Measures for All Possible Thresholds}
\usage{
thresholds(truth, prob, positive, measures = c("tpr", "tnr"))
}
\arguments{
\item{truth}{:: \code{factor()}\cr
True (observed) labels.
Must have the exactly same two levels and the same length as \code{response}.}

\item{prob}{:: \code{numeric()}\cr
Predicted probability for positive class.
Must have exactly same length as \code{truth}.}

\item{positive}{:: \code{character(1)}\cr
Name of the positive class.}
}
\value{
(\code{matrix()}). First column are the distinct probability values, additional
columns store the calculated performance values, one column for each measure.
}
\description{
For each value in probability \code{prob}, a confusion matrix is build and the
all requested measures are calculated.
}
\examples{
truth = factor(c("a", "a", "a", "b"))
prob = c(.6, .7, .1, .4)
thresholds(truth, prob, "a")
}
