% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif_tpr.R
\name{tpr}
\alias{tpr}
\title{True Positive Rate}
\usage{
tpr(truth, response, positive, na_value = NaN)
}
\arguments{
\item{truth}{:: \code{factor()}\cr
True (observed) labels.
Must have the same length as \code{response}.}

\item{response}{:: \code{factor()}\cr
Predicted response labels.
Must have the same length as \code{truth}.}

\item{positive}{:: \code{character(1)}\cr
Name of the positive class.}

\item{na_value}{:: \code{numeric(1)}\cr
Value that should be returned if the measure is not defined for the input
(as described in the note). Default is \code{NaN}.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Definition: \deqn{
   \frac{TP}{TP + FN}
}{
 TP / (TP + FN)
}
}
\section{Meta Information}{

\itemize{
\item Minimum value: 0
\item Maximum value: 1
\item Minimize: FALSE
}
}

