% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_auc.R
\name{auc}
\alias{auc}
\title{Area Under the ROC Curve}
\usage{
auc(truth, prob, positive)
}
\arguments{
\item{truth}{:: \code{factor()}\cr
True (observed) labels.
Must have exactly same two levels and the same length as \code{response}.}

\item{prob}{:: \code{numeric()}\cr
Predicted probability for positive class.
Must have exactly same length as \code{truth}.}

\item{positive}{:: \code{character(1)}\cr
Name of the positive class.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Computes the area under the Receiver Operator Characteristic (ROC) curve.
The AUC can be interpreted as the probability that a randomly chosen positive observation
has a higher predicted probability than a randomly chosen negative observation.
}
\section{Meta Information}{

\itemize{
\item Minimum value: 0
\item Maximum value: 1
\item Minimize: FALSE
}
}

\seealso{
Other Binary Classification Measures: \code{\link{dor}},
  \code{\link{fbeta_score}}, \code{\link{fdr}},
  \code{\link{fnr}}, \code{\link{fn}}, \code{\link{fomr}},
  \code{\link{fpr}}, \code{\link{fp}}, \code{\link{mcc}},
  \code{\link{npv}}, \code{\link{ppv}}, \code{\link{tnr}},
  \code{\link{tn}}, \code{\link{tpr}}, \code{\link{tp}}

Other Classification Measures: \code{\link{acc}},
  \code{\link{ce}}, \code{\link{dor}},
  \code{\link{fbeta_score}}, \code{\link{fdr}},
  \code{\link{fnr}}, \code{\link{fn}}, \code{\link{fomr}},
  \code{\link{fpr}}, \code{\link{fp}}, \code{\link{mcc}},
  \code{\link{npv}}, \code{\link{ppv}}, \code{\link{tnr}},
  \code{\link{tn}}, \code{\link{tpr}}, \code{\link{tp}}
}
\concept{Binary Classification Measures}
\concept{Classification Measures}
