% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_fp.R
\name{fp}
\alias{fp}
\title{False Positives}
\usage{
fp(truth, response, positive, ...)
}
\arguments{
\item{truth}{:: \code{factor()}\cr
True (observed) labels.
Must have exactly same two levels and the same length as \code{response}.}

\item{response}{:: \code{factor()}\cr
Predicted response labels.
Must have exactly same two levels and the same length as \code{truth}.}

\item{positive}{:: \code{character(1)}\cr
Name of the positive class.}

\item{...}{:: any\cr
Additional arguments. Currently ignored.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Classification measure counting the false positives (type 1 error), i.e. the number of
predictions indicating a positive class label while in fact it is negative.
}
\section{Meta Information}{

\itemize{
\item Minimum value: 0
\item Maximum value: Inf
\item Minimize: TRUE
}
}

\references{
\url{https://en.wikipedia.org/wiki/Template:DiagnosticTesting_Diagram}
}
\seealso{
Other Binary Classification Measures: \code{\link{auc}},
  \code{\link{dor}}, \code{\link{fbeta}},
  \code{\link{fdr}}, \code{\link{fnr}}, \code{\link{fn}},
  \code{\link{fomr}}, \code{\link{fpr}}, \code{\link{mcc}},
  \code{\link{npv}}, \code{\link{ppv}}, \code{\link{tnr}},
  \code{\link{tn}}, \code{\link{tpr}}, \code{\link{tp}}
}
\concept{Binary Classification Measures}
\concept{binary_classification_measure}
