% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classif_logloss.R
\name{logloss}
\alias{logloss}
\title{Log Loss}
\usage{
logloss(truth, prob, eps = 1e-15)
}
\arguments{
\item{truth}{:: \code{factor()}\cr
True (observed) labels.
Must have the same levels and length as \code{response}.}

\item{prob}{:: \code{matrix()}\cr
Matrix of predicted probabilities, each column is a vector of probabilities for a
specific class label.
Columns must be named with levels of \code{truth}.}

\item{eps}{:: \code{numeric(1)}\cr
Probabilities are clipped to \code{max(eps, min(1 - eps, p))}, otherwise the measure would
be undefined for probabilities \code{p = 0} and \code{p = 1}.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Classification measure defined as \deqn{
  -\frac{1}{n} \sum_{i=1}^n \log \left(  p_i \right )
}{
  -mean(log(p))
}
where \eqn{p_i}{p} is the probability for the true class of observation \eqn{i}.
}
\section{Meta Information}{

\itemize{
\item Minimum value: 0
\item Maximum value: Inf
\item Minimize: TRUE
}
}

\seealso{
Other Classification Measures: \code{\link{acc}},
  \code{\link{ber}}, \code{\link{ce}}
}
\concept{Classification Measures}
\concept{classification_measure}
