% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr_mae.R
\name{mae}
\alias{mae}
\title{Mean Absolute Errors}
\usage{
mae(truth, response, ...)
}
\arguments{
\item{truth}{:: \code{numeric()}\cr
True (observed) values.
Must have the same length as \code{response}.}

\item{response}{:: \code{numeric()}\cr
Predicted response values.
Must have the same length as \code{truth}.}

\item{...}{:: any\cr
Additional arguments. Currently ignored.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Regression measure defined as \deqn{
  \frac{1}{n} \sum_{i=1}^n \left| t_i - r_i \right|.
}{
  mean(abs(t - r)).
}
}
\section{Meta Information}{

\itemize{
\item Minimum value: 0
\item Maximum value: Inf
\item Minimize: TRUE
}
}

\seealso{
Other Regression Measures: \code{\link{bias}},
  \code{\link{kendalls_tau}}, \code{\link{mape}},
  \code{\link{maxae}}, \code{\link{maxse}},
  \code{\link{medae}}, \code{\link{medse}},
  \code{\link{mse}}, \code{\link{msle}},
  \code{\link{percent_bias}}, \code{\link{rae}},
  \code{\link{rmse}}, \code{\link{rmsle}},
  \code{\link{rrse}}, \code{\link{rse}}, \code{\link{rsq}},
  \code{\link{sae}}, \code{\link{smape}},
  \code{\link{spearmans_rho}}, \code{\link{sse}}
}
\concept{Regression Measures}
