% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measures.R
\docType{data}
\name{measures}
\alias{measures}
\title{Measure Registry}
\format{
An object of class \code{environment} of length 47.
}
\usage{
measures
}
\description{
The \code{\link[=environment]{environment()}} \code{measures} keeps track of all measures in this package.
Stores meta information about measures, such as minimum, maximum of if the
measure must be minimized or maximized.
}
\examples{
names(measures)
measures$tpr
}
\keyword{datasets}
