% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3measures-package}
\alias{mlr3measures}
\alias{mlr3measures-package}
\title{mlr3measures: Performance Measures for 'mlr3'}
\description{
Implements multiple performance measures for supervised learning. Includes over 40 measures for regression and classification. Additionally, meta information about the performance measures can be queried, e.g. what the best and worst possible performances scores are.
}
\seealso{
Useful links:
\itemize{
  \item \url{https:///mlr3measures.mlr-org.com}
  \item \url{https://github.com/mlr-org/mlr3measures}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3measures/issues}
}

}
\author{
\strong{Maintainer}: Marc Becker \email{marcbecker@posteo.de} (\href{https://orcid.org/0000-0002-8115-0400}{ORCID})

Authors:
\itemize{
  \item Michel Lang \email{michellang@gmail.com} (\href{https://orcid.org/0000-0001-9754-0393}{ORCID})
}

Other contributors:
\itemize{
  \item Martin Binder \email{mlr.developer@mb706.com} [contributor]
}

}
