% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr_rae.R
\name{rae}
\alias{rae}
\title{Relative Absolute Error}
\usage{
rae(truth, response, na_value = NaN)
}
\arguments{
\item{truth}{:: \code{numeric()}\cr
True (observed) values.
Must have the same length as \code{response}.}

\item{response}{:: \code{numeric()}\cr
Predicted response values.
Must have the same length as \code{truth}.}

\item{na_value}{:: \code{numeric(1)}\cr
Value that should be returned if the measure is not defined for the input
(as described in the note). Default is \code{NaN}.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Definition:
\deqn{
  \frac{\sum_{i=1}^n \left| t_i - r_i \right|}{\sum_{i=1}^n \left| t_i - \bar{t} \right|}.
}{
  sum((t - r)^2) / sum((t - mean(t))^2)
}
}
\note{
This measure is undefined for constant \eqn{t}.
}
\section{Meta Information}{

\itemize{
\item Minimum value: 0
\item Maximum value: Inf
\item Minimize: TRUE
}
}

