% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regr_spearmans_rho.R
\name{spearmans_rho}
\alias{spearmans_rho}
\title{Spearman's rho}
\usage{
spearmans_rho(truth, response)
}
\arguments{
\item{truth}{:: \code{numeric()}\cr
True (observed) values.
Must have the same length as \code{response}.}

\item{response}{:: \code{numeric()}\cr
Predicted response values.
Must have the same length as \code{truth}.}
}
\value{
Performance value as \code{numeric(1)}.
}
\description{
Regression measures defined as Spearman's rank correlation coefficient between truth and response.
Calls \code{\link[stats:cor]{stats::cor()}} with \code{method} set to \code{"spearman"}.
}
\section{Meta Information}{

\itemize{
\item Minimum value: -1
\item Maximum value: 1
\item Minimize: FALSE
}
}

\references{
\cite{rosset_2006}
}
\seealso{
Other Regression Measures: \code{\link{kendalls_tau}},
  \code{\link{mae}}, \code{\link{mape}},
  \code{\link{maxae}}, \code{\link{maxse}},
  \code{\link{medae}}, \code{\link{medse}},
  \code{\link{mse}}, \code{\link{msle}}, \code{\link{rae}},
  \code{\link{rmse}}, \code{\link{rmsle}},
  \code{\link{rrse}}, \code{\link{rse}}, \code{\link{rsq}},
  \code{\link{sae}}, \code{\link{smape}}, \code{\link{sse}}
}
\concept{Regression Measures}
