% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholds.R
\name{thresholds}
\alias{thresholds}
\title{Calculate Measures for All Possible Thresholds}
\usage{
thresholds(truth, prob, positive, measures = c("tpr", "tnr"))
}
\arguments{
\item{truth}{:: \code{factor()}\cr
True (observed) labels.
Must have the exactly same two levels and the same length as \code{response}.}

\item{prob}{:: \code{numeric()}\cr
Predicted probability for positive class.
Must have exactly same length as \code{truth}.}

\item{positive}{:: \code{character(1)}\cr
Name of the positive class.}
}
\value{
(\code{matrix()}). First column are the distinct probability values, additional
columns store the calculated performance values, one column for each measure.
}
\description{
For each value in probability \code{prob}, a confusion matrix is build and the
all requested measures are calculated.
}
\examples{
N = 200
truth = factor(sample(c("a", "b"), N, replace = TRUE), levels = c("a", "b"))
prob = runif(N)
th = thresholds(truth, prob, "a")

# poor mans roc curve
plot(th[, 2], 1 - th[, 3], type = "l")
abline(0, 1)
}
